requirements.txt.txt
tensorflow
streamlit
streamlit-lottie
Pillow
numpy
gdown
